# Conflicts:
#	bitbucket-pipelines.yml

+++ software/program/ESP_Reciever/include/TDMA.h
@@ -5,7 +5,7 @@
 #include "Custom_Lora.h"
 #include "stdint.h"
 #include <RadioLib.h>
-
+#include "mbedtls/aes.h"
 
 // Cấu hình cho scheduler beacon
 struct TDMA_BeaconConfig {
@@ -17,7 +17,7 @@ struct TDMA_BeaconConfig {
 };
 
 struct __attribute__((packed)) TDMABeacon {
-  uint8_t   sync;               // = 0xAA
+  uint16_t  sync;               // = 0xAA
   uint16_t  frame_id;           // ID frame hiện tại (do master sinh)
   uint32_t  beacon_timestamp;   // millis() tại master khi gửi
   uint16_t  slot_len_ms;        // độ dài 1 slot
